Refactor authentication logic to derive hashed passwords and compute responses; streamline WebSocket message handling
